CCLE-3855 - fixing 'instance_delete' bug
this prevents the main SITE alert block from being deleted
when a course with an alert block instance is deleted.
previously, assumption was made that $COURSE object 
corresponded to course where alert block instance lived,
but in the context of a course deletion, $COURSE is actually
the main SITE.  this incorrect assumption led to unwanted 
deletion of main SITE alert block.

this fix uses the block's instanceid to correctly get the courseid
needed to delete the block's DB content