CCLE-3034 - grouping crosslist bug
this bug was happening because the grouping name 
generating function was not returning a unique 
name for 'course_lecture'.  The result of this is that a 
grouping of all groups in the set is created and is skipped --
by design -- and thus this bug

I've added the 'lecnum' field so that a unique grouping
name is generated for the different lectures.  For the test
case:
  LIFESCI1-1
  LIFESCI1-2
The groupings generated are
  LIFESCI 1 1 DEM
  LIFESCI 1 2 DEM
With their respective groups in their respective sets